working large pages

+++ kernel/virtual_MM.h
@@ -7,7 +7,7 @@
 
 uint32_t newCPTD(unsigned char domain,uint32_t base_address);
 
-uint32_t newLPD(unsigned char c,unsigned char b,unsigned char ap,uint16_t base_address);
+uint32_t newLPD(unsigned char c,unsigned char b,unsigned char ap,uint32_t base_address);
